Check shaman not only for repo but for build complete
The package repo might be done, but the build must be totally
complete for the container to be present.  For a specific build
(in global variables, for now), demand that the build be complete
before claiming packages exist.  This allows suite jobs to
fail or --newest to continue searching when the container isn't
done.

Signed-off-by: Dan Mick <dmick@redhat.com>