doc/dev: update essentials.rst
This commit changes the noun "backup" to the
intended verb "back up" (this is one of my
technical documentation pet peeves and I couldn't
resist), and it also removes the verb "to nuke"
and replaces it with something less poetic but
more likely to be discoverable in a dictionary by
someone whose English is weak. It also adds a
missing colon.

Signed-off-by: Zac Dover <zac.dover@gmail.com>