doc/start: update prompts in documenting-ceph.rst
This commit adds the ".. prompt:: bash $" directive
to the commands in "documenting-ceph.rst" so that
the permissions level of each command is evident
(through the presentation of the $ or the #), and
so that when the reader cuts and pastes the commands
from the documentation, they don't have to delete
the prompt (because that's how this directive
renders).

This commit also instructs the reader to add the
directory containing "build-doc" and
"serve-doc" to the PATH environment variable
in order to run the commands.

Signed-off-by: Zac Dover <zac.dover@gmail.com>
Co-authored-by: Kefu Chai <tchaikov@gmail.com>