server: invoke schdule_request() while requests are available
The problem is that the further request is not fetched when
the schedule_request() executes the submit_request() and
sched_ahead_when is set to TimeZero. It implies that run_sched_ahead
thread can't wake up by itself. This commit allows that
the schedule_request() is called while schedule_request returns
NextReqType that indicates us whether schedule or not.

Signed-off-by: Yongseok Oh <yongseok.oh@linecorp.com>