osd: fix clone size mismatch when dedup object is evicted
When evicting deduped object which is cloned,
we try to shrink its size to zero if all chunks are in chunk_map.
However, we did not set clone_size to zero, resulting in
clone size mismatch.

Specifically, build_push_op appends zero data with reference to
copy_set set by clone_size, but actual size in oi.size is zero.

Signed-off-by: Myoungwon Oh <myoungwon.oh@samsung.com>