client: ensure we take Fs caps when querying dirstat vxattrs
Dirstats values are only updated when Fs caps are held, so ensure that
we request those caps in the _getattr call when querying for them.

Drop the old XATTR_NAME_CEPH helper and rename XATTR_NAME_CEPH2 to
XATTR_NAME_CEPH. Add a new VXATTR_DIRSTAT field and add that to all of
the dirstats vxattrs.

Fixes: https://tracker.ceph.com/issues/48313
Signed-off-by: Jeff Layton <jlayton@redhat.com>