cls/rgw: update new current bi entries epoch
Fixes: https://tracker.ceph.com/issues/48268

When unlinking current instance, update the new current instance with
the new versioned epoch. This was broken before, because we failed to
initialize the new current entry.
Also, when linking entry, check if olh already points to it when
determining if it's current. This works around a case where instance epoch
is incorrectly lower than the current epoch.

Signed-off-by: Yehuda Sadeh <yehuda@redhat.com>