make-dist: run with -x
This script already produces lots of output, but the output is not easy to
understand because it's produced by a bunch of shell commands (i.e. the commands
that make up the script "make-dist") which aren't shown in the output. The cost
of the extra output produced by adding "-x" is outweighed by the benefit of
being able to identify which shell commands produced which parts of the output.

In other words, this makes it easier to debug the script when it breaks.

Signed-off-by: Nathan Cutler <ncutler@suse.com>