rgw: update_olh() needs to use zones_trace
A zone might need apply a change that would cause cascading changes,
e.g., when creating a new delete marker with same instance as existing
object (can happen in version-suspended bucket), the olh log will
reflect object removal. If we don't pass the zones_trace, the object
removal might be applied erroneously on the delete marker in the
originating zone.

Signed-off-by: Yehuda Sadeh <yehuda@redhat.com>