mgr: do not migrate conf from config-key store to new-style conf
since all module options are using the new-style config framework.
the migration is offered for the use case of upgrade from luminous to mimic,
since pacific can only be upgraded from octopus. the mimic monitors are alreay
able to populate the configurations to mgr, not to mention the octopus
monitors, so there is no need to migrate the options stored in config-key
store anymore.

Signed-off-by: Kefu Chai <kchai@redhat.com>