cephadm: use 2xx error codes instead 5xx
When a thread fails we still need to see content from the
other threads. When a 5xx response is received the payload
is absent - so by using 200, 204 and 206 we can understand
thread health from the status code, saving a 500 response
to indicate all threads are dead!

In addition when a thread dies the data payload in the cache
is erased to prevent stale information being returned.

Signed-off-by: Paul Cuzner <pcuzner@redhat.com>