ceph-volume: prepare's return value should depend on disk status
When prepare subcommand is executed with a device that is already a Ceph
device, check if the device is used how it is described in the command
and, accordingly, print a message and exit with a return value.

This affects create subcommand in very same way since prepare
functionality is utilized internally.

Fixes: https://tracker.ceph.com/issues/47584
Signed-off-by: Rishabh Dave <ridave@redhat.com>