osd: osdmap: Do not increase compatv when using stretch mode
This is analagous to my monmap fix in 2e3643647bfbe955b54c62c8aaf114744dedb86e,
but for the OSDMap. I was previously trying to be safe by changing compat_v
when using stretch details. Unfortunately, older clients are trying to
interpret this data (even though its purpose is to be a cluster-only data
region! See https://tracker.ceph.com/issues/48489), and crashing as a result.

Because enabling stretch mode is blocked by all OSDs understanding it, and
OSDs which do not understand it are not allowed to boot once it's on,
we can simply remove the compat_v change.

This does leave the miniscule risk that somebody will enable stretch mode
and run with it, then turn it off, an old OSD will *then* connect to the
cluster, and maybe some bad peering choices will happen -- but I don't
see a lot of other choices and this scenario is so unlikely I don't
think it's worth worrying about.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>