ceph-volume: util.disk.is_locked_raw_device refactor
The current implementation of `is_locked_raw_device()` tries to open and
acquire an exclusive lock on the device in order to detect if the device
is locked. This might trigger udev events and cause race condition.

using `lslocks` which is present on all Linux distribution as part of
`util-linux` reads `/proc/locks` and doesn't trigger udev events.

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>