mgr/dashboard: Fixing dashboard logs e2e test
Yesterday there were some changes went in which disables the logging of config set and config-key set to to be logged in mgr audit logs. Dashboard has an e2e test which checks for this config set and broke the current e2e jenkins job. This commit removes that certain test to fix the jenkins job.

Fixes: https://tracker.ceph.com/issues/48623
Signed-off-by: Nizamudeen A <nia@redhat.com>