crimson/osd: create the collected push ops use shared_ptr
do_with() uses a unique_ptr for keeping the captured variables alive. so
it is practically the same thing in concept. but `do_with()` is more
distracting and difficult to reason about, it increases the indent level
also.

so use a lw_shared_ptr() for collecting the push ops map.

Signed-off-by: Kefu Chai <kchai@redhat.com>