crimson/osd: do not use the parallel_for_each() in prep_push()
the caller of prep_push() also loops in target pg shards, so there is no
need to do the loop in prep_push(). in this change, prep_push() just
takes care of a single pg_shard for a single oid instead of collecting
all push ops for all pg_shards to be recovered.

Signed-off-by: Kefu Chai <kchai@redhat.com>