cephadm: make host add failure message more friendly
When a host add fails, the output was showing passed
and failed checks. This patch uses an ERROR prefix for
all failed checks, and then filters on them when showing
the result, so the admin sees only the failed items. The
text has also been tidied up to remove the [] within a
[] syntax

Signed-off-by: Paul Cuzner <pcuzner@redhat.com>