common/ConfUtils: pass string_view instead of string
pass string_view as the section_name when reading options from a
ConfFile. as there is no need to create a temporary string object
if the section name is a `const char*` literal.

also, update the caller site in `common/config.cc` accordingly as a
cleanup. as it's not necessary to pass a `const char*` for the section
name even without this change.

Signed-off-by: Kefu Chai <kchai@redhat.com>