mgr/dashboard: Fix for datatable item not showing details after getting selected
**Regression in nautilus**

Datatable items are not showing the details even if an item in the list is selected. This is happening because of this backport (https://github.com/ceph/ceph/pull/37756/files) which backports the line this.selection.selected = $event['selected']; but this feature was not implemented in the nautilus branch originally.

Fixes: https://tracker.ceph.com/issues/48796
Signed-off-by: Nizamudeen A <nia@redhat.com>