rpm: ceph-mgr-dashboard recommends python3-saml on SUSE
This commit is not cherry-picked from master because it is fixing an incomplete
backport.

The master commit

    d8d3b33633c0a995aef034fdabc55d47c3872566
    rpm: add python3-saml as install dependency

added a line

    Recommends:     python%{python3_pkgversion}-python3-saml

which was intentionally omitted from the nautilus cherry-pick because this
package (python3-python3-saml) was not available in SLE-15-SP1/SES6. In
the meantime, the package has been made available there, so the line can
now be backported to nautilus.

Fixes: 59f704bfd94728f350e5a6fc6f54416fb5534eeb
Signed-off-by: Nathan Cutler <ncutler@suse.com>