mgr/cephadm: Use our image for remote tool execution
Fixes: https://tracker.ceph.com/issues/48870

This deals with a cephadm upgrade issue:

1. user calls `ceph orch upgrade`
2. mgr/cephadm calls `ceph orch config set mgr.x container_image <new-container>`
3. standby mgr gets upgraded
4. mgr failover to new mgr
5. mgr/cephadm calls `_refresh_host_devices`
6. `_refresh_host_devices` calls` ceph orch config get osd container_image`.
  But this returns the old image
7. `_refresh_host_devices` calls `ceph-volume ... --filter-for-batch`
  with an image that doesn't support `filter-for-batch`

The idea is to _always_ call tools with the same container image as
run by the active mgr. This way we don't need to deal with CLI changes at all.
Downside is: tools need to be able to cope with being called
with alternating versions on the same host.

Signed-off-by: Sebastian Wagner <sebastian.wagner@suse.com>