pybind/mgr/mgr_module: add poll param to CLICommand
and add it to the output of dump_cmd() call, if "poll" is missing
in the output of dump_cmd() call, ceph CLI consider the it as False.

but it'd be better to be explicit. so this field is always set.

Signed-off-by: Kefu Chai <kchai@redhat.com>
(cherry picked from commit 7a2e460d34dc5f7e2124bcbedd527d63fa467017)