Revert "*: remove legacy ceph_volume_client.py library"
This reverts commit a3db265ad5b3ff899d8b71164a7f9ea5a426b19e.
The ceph_volume_client is being used by manila and it's
not fully equipped to use mgr/volumes and expects to use
ceph_volume_client for one more release. Hence reverting
this commit for pacific release.

Note that this is a pacific only patch and hence no corresponding
master patch is available.

Fixes: https://tracker.ceph.com/issues/48923
Signed-off-by: Kotresh HR <khiremat@redhat.com>